fix coding problems and add notes
