+++ src/Interpreters/ExpressionAnalyzer.cpp
@@ -279,25 +279,22 @@ void tryTranslateToParametricAggregateFunction(
     }
     else if (lower_name == "group_concat")
     {
+        /// Translate `group_concat(expression, delimiter, limit)` to `group_concat(delimiter, limit)(expression)`
         if (arguments.size() > 3 || arguments.size() < 1)
         {
             throw Exception(ErrorCodes::NUMBER_OF_ARGUMENTS_DOESNT_MATCH,
                 "Incorrect number of parameters for aggregate function {}, should be 0, 1 or 2, got: {}", node->name, parameters.size());
         }
-        else if (arguments.size() == 3)
+
+        if (arguments.size() > 1)
         {
             ASTPtr expression_list = std::make_shared<ASTExpressionList>();
-            for (size_t i = 1; i < 3; i++)
+            for (size_t i = 1; i < arguments.size(); i++)
                 expression_list->children.push_back(arguments[i]);
 
             parameters = getAggregateFunctionParametersArray(expression_list, "", context);
         }
-        else if (arguments.size() == 2)
-        {
-            ASTPtr expression_list = std::make_shared<ASTExpressionList>();
-            expression_list->children.push_back(arguments[1]);
-            parameters = getAggregateFunctionParametersArray(expression_list, "", context);
-        }
+
         argument_names = {argument_names[0]};
         types = {types[0]};
     }
