+++ tests/queries_ported/0_stateless/03156_group_concat.reference
@@ -10,9 +10,10 @@ abc,a,makson95
 [1,2,3],[993,986,979,972]
 \N
 0
-951239595123123
-abc,a,makson95,abc,abc,a,a,makson95,makson95
-[1,2,3][993,986,979,972][][1,2,3][1,2,3][993,986,979,972][993,986,979,972][][]
+0
+951239512395123
+abc,a,makson95,abc,a,makson95,abc,a,makson95
+[1,2,3][993,986,979,972][][1,2,3][993,986,979,972][][1,2,3][993,986,979,972][]
 488890
 0
 488890
