+++ tests/queries_ported/0_stateless/03235_groupArray_returns_string.reference
@@ -1,0 +1,2 @@
+0
+true
